refactoring bcc::core
Splitting the new structs in the recent refactoring to higher-level
modules. This allows bcc::core to retain an internal representation
of concrete loaded/attached probes, while the higher-level modules
define the public interface for constructing and attaching probes.
