+++ examples/contextswitch.rs
@@ -2,8 +2,8 @@
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
-use bcc::core::{PerfEventProbe, BPF};
-use bcc::perf::{Event, SoftwareEvent};
+use bcc::{PerfEvent, BPF};
+use bcc::perf_event::{Event, SoftwareEvent};
 use bcc::BccError;
 use clap::{App, Arg};
 
@@ -95,7 +95,7 @@ fn do_main(runnable: Arc<AtomicBool>) -> Result<(), BccError> {
     }
 
     let mut bpf = BPF::new(&code)?;
-    PerfEventProbe::new()
+    PerfEvent::new()
         .name("do_count")
         .event(Event::Software(SoftwareEvent::ContextSwitches))
         .sample_period(sample_period)
